+++ src/main/groovy/com/github/eerohele/SaxonPlugin.groovy
@@ -1,4 +1,4 @@
-package com.github.Lunatix01
+package com.github.eerohele
 
 import org.gradle.api.Project
 import org.gradle.api.Plugin
