+++ src/main/groovy/com/github/eerohele/XsltTransformation.groovy
@@ -1,4 +1,4 @@
-package com.github.Lunatix01
+package com.github.eerohele
 
 import net.sf.saxon.Transform
 import org.gradle.workers.WorkAction
